+++ src/fftpack/rfftb.f90
@@ -94,7 +94,7 @@ SUBROUTINE RFFTB(N,R,Wsave)
   INTEGER :: ifac(15)
   !***FIRST EXECUTABLE STATEMENT  RFFTB
   IF ( N==1 ) RETURN
-  ifac = Wsave(2*N+1:2*N+15)
+  ifac = INT( Wsave(2*N+1:2*N+15) )
   CALL RFFTB1(N,R,Wsave,Wsave(N+1),ifac)
   Wsave(2*N+1:2*N+15) = ifac
 END SUBROUTINE RFFTB
